Only prepend MuackPrepended if really needed:
If all the prepended modules do not have the method we're mocking,
there's no need to prepend another module, because we can simply
define the method on the singleton class.

We only need to prepend another module to override the methods in
the prepended modules. Bottom line: we cannot touch the prepended
modules, because that would affect other modules/class using those
modules.

This also means we need to memorize which target we picked to change,
and then we can safely cleanup the target, which can be in the middle
of the ancestor chain when there are prepended modules yet we're
touching the singleton class not the internal prepended module.